Move loss funtion to standalone file
