+++ recognition/s4574278/dataset.py
@@ -56,14 +56,7 @@ class IsicDataSet(Dataset):
 
 
 ##########################################################
-# Loader
-##########################################################
-
-
-
-
-##########################################################
-# Helpers
+# Transforms
 ##########################################################
